+++ editor.planx.uk/src/@planx/components/Checklist/Public/tests/Public.test.tsx
@@ -8,124 +8,8 @@ import { setup } from "testUtils";
 import { vi } from "vitest";
 import { axe } from "vitest-axe";
 
-import { Option } from "../../shared";
-import { Group } from "../model";
-import Checklist, { ChecklistLayout } from "./Public";
-
-const options: {
-  [key in ChecklistLayout]?: Array<Option>;
-} = {
-  [ChecklistLayout.Basic]: [
-    {
-      id: "flat_id",
-      data: {
-        text: "Flat",
-      },
-    },
-    {
-      id: "caravan_id",
-      data: {
-        text: "Caravan",
-      },
-    },
-    {
-      id: "house_id",
-      data: {
-        text: "House",
-      },
-    },
-    {
-      id: "spaceship_id",
-      data: {
-        text: "Spaceship",
-      },
-    },
-  ],
-  [ChecklistLayout.Images]: [
-    {
-      id: "flat_id",
-      data: {
-        text: "Flat",
-        img: "flat.jpg",
-      },
-    },
-    {
-      id: "caravan_id",
-      data: {
-        text: "Caravan",
-        img: "caravan.jpg",
-      },
-    },
-    {
-      id: "house_id",
-      data: {
-        text: "House",
-        img: "house.jpg",
-      },
-    },
-    {
-      id: "spaceship_id",
-      data: {
-        text: "Spaceship",
-        img: "spaceship.jpg",
-      },
-    },
-  ],
-};
-
-const groupedOptions: Array<Group<Option>> = [
-  {
-    title: "Section 1",
-    children: [
-      {
-        id: "S1_Option1",
-        data: {
-          text: "S1 Option1",
-        },
-      },
-      {
-        id: "S1_Option2",
-        data: {
-          text: "S1 Option2",
-        },
-      },
-    ],
-  },
-  {
-    title: "Section 2",
-    children: [
-      {
-        id: "S2_Option1",
-        data: {
-          text: "S2 Option1",
-        },
-      },
-      {
-        id: "S2_Option2",
-        data: {
-          text: "S2 Option2",
-        },
-      },
-    ],
-  },
-  {
-    title: "Section 3",
-    children: [
-      {
-        id: "S3_Option1",
-        data: {
-          text: "S3 Option1",
-        },
-      },
-      {
-        id: "S3_Option2",
-        data: {
-          text: "S3 Option2",
-        },
-      },
-    ],
-  },
-];
+import Checklist, { ChecklistLayout } from "../Public";
+import { groupedOptions, options } from "./mockOptions";
 
 describe("Checklist Component - Grouped Layout", () => {
   it("answers are submitted in order they were supplied", async () => {
@@ -201,7 +85,7 @@ describe("Checklist Component - Grouped Layout", () => {
         groupedOptions={groupedOptions}
       />,
     );
-    const [section1Button, section2Button, section3Button, _continueButton] =
+    const [section1Button, section2Button, section3Button] =
       screen.getAllByRole("button");
 
     // All accordion sections begin collapsed
